REFACTOR: RecruitType package 이동
