+++ recruit-jogbo-common/src/main/java/univ/study/recruitjogbo/post/recruitType/RecruitType.java
@@ -1,4 +1,4 @@
-package univ.study.recruitjogbo.post;
+package univ.study.recruitjogbo.post.recruitType;
 
 import lombok.AccessLevel;
 import lombok.Getter;
